ai_agent/integration.py
```python
import requests
from .local_computer_control import executeCommand
from .task_automation import scheduleTask
from .chat_interaction import respondToChat
from .web_browser_integration import navigateWebPage, performWebSearch
from .internet_access import retrieveInformation
from .voice_control import processVoiceCommand
from .personalization_learning import learnFromInteraction
from .security_privacy import protectData

class Integration:
    def __init__(self, api_url):
        self.api_url = api_url

    def extendFunctionality(self, function_name, *args, **kwargs):
        if function_name == 'executeCommand':
            return executeCommand(*args, **kwargs)
        elif function_name == 'scheduleTask':
            return scheduleTask(*args, **kwargs)
        elif function_name == 'respondToChat':
            return respondToChat(*args, **kwargs)
        elif function_name == 'navigateWebPage':
            return navigateWebPage(*args, **kwargs)
        elif function_name == 'performWebSearch':
            return performWebSearch(*args, **kwargs)
        elif function_name == 'retrieveInformation':
            return retrieveInformation(*args, **kwargs)
        elif function_name == 'processVoiceCommand':
            return processVoiceCommand(*args, **kwargs)
        elif function_name == 'learnFromInteraction':
            return learnFromInteraction(*args, **kwargs)
        elif function_name == 'protectData':
            return protectData(*args, **kwargs)
        else:
            raise ValueError(f"Function {function_name} not found")

    def integrateWithExternalApp(self, app_name, data):
        response = requests.post(f"{self.api_url}/{app_name}", data=data)
        return response.json()
```