ai_agent/tests/test_integration.py
```python
import unittest
from ai_agent import integration

class TestIntegration(unittest.TestCase):

    def setUp(self):
        self.integration = integration.Integration()

    def test_open_api(self):
        result = self.integration.open_api()
        self.assertTrue(result)

    def test_extend_functionality(self):
        result = self.integration.extend_functionality()
        self.assertTrue(result)

    def test_integrate_with_applications(self):
        result = self.integration.integrate_with_applications()
        self.assertTrue(result)

if __name__ == '__main__':
    unittest.main()
```